# web/openapi.yaml
openapi: 3.0.3
info:
  title: work-tracker API
  description: Automatically generated OpenAPI spec from Postman collection "work-tracker"
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
    description: Default local server
paths:
...
  /api/work/options:
    get:
      summary: Get task categories, projects and the work table setting
      description: Categories and projects include inactive ones, so old entries can still show their names.
      operationId: getWorkOptions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOptionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/work/members:
    get:
      summary: List members for the everyone's work table
      description: Allowed for admins, and for everyone when `allowViewAll` is on.
      operationId: listWorkMembers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkMembersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/work-records:
    get:
      summary: List everyone's work records
      description: Every record dated from..to (the frontend asks for one week), newest date first. Allowed for admins, and for everyone when `allowViewAll` is on.
      operationId: listAllWorkRecords
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - name: account
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkRecordListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/me/work-records:
    get:
      summary: List my work records
      description: Every record dated from..to (the frontend asks for one week), newest date first.
      operationId: listMyWorkRecords
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkRecordListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Add a work record
      operationId: createMyWorkRecord
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveWorkEntryRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkRecordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/me/work-records/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      summary: Replace one of my work records
      operationId: updateMyWorkRecord
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveWorkEntryRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkRecordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete one of my work records
      operationId: deleteMyWorkRecord
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/me/todos:
    get:
      summary: List my todos
      description: Sorted by date, earliest first.
      operationId: listMyTodos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodosResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Add a todo
      operationId: createMyTodo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveWorkEntryRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/me/todos/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      summary: Replace one of my todos
      operationId: updateMyTodo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveWorkEntryRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete one of my todos
      operationId: deleteMyTodo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/me/todos/{id}/complete:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      summary: Complete a todo
      description: Deletes the todo and adds it as a work record dated `date` (the client's today). Answers 400 and keeps the todo when the record would lack a category or hours.
      operationId: completeMyTodo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTodoRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkRecordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/me/week-summary:
    get:
      summary: Summarize my week
      description: >-
        Records and hours of the 7 days starting at `from`, and the hours to
        log: 8 per workday, where workdays are Monday-Friday minus holidays plus
        makeup workdays from the holiday calendar.
      operationId: getWeekSummary
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          description: First day of the week (a Monday in the UI), YYYY-MM-DD.
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeekSummaryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    From:
      name: from
      in: query
      required: true
      description: Inclusive start date (YYYY-MM-DD).
      schema:
        type: string
        format: date
    To:
      name: to
      in: query
      required: true
      description: Inclusive end date (YYYY-MM-DD), not before `from`.
      schema:
        type: string
        format: date
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
  schemas:
...
    CategoryColor:
      type: string
      description: Palette color name of a task category. Projects have no color. New categories get the least used color.
      enum:
        - blue
        - sky
        - teal
        - green
        - lime
        - amber
        - orange
        - red
        - pink
        - purple
    WorkOption:
      type: object
      required:
        - id
        - name
        - active
      properties:
        id:
          type: string
        name:
          type: string
        active:
          type: boolean
          description: Inactive options are hidden from new entries but still name old ones.
        color:
          $ref: '#/components/schemas/CategoryColor'
    WorkOptionsResponse:
      type: object
      required:
        - message
        - categories
        - projects
        - allowViewAll
      properties:
        message:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/WorkOption'
        projects:
          type: array
          items:
            $ref: '#/components/schemas/WorkOption'
        allowViewAll:
          type: boolean
...
    SaveWorkEntryRequest:
      type: object
      description: >-
        Work records require categoryId and hours; todos only require date and
        description. projectId is always optional.
      required:
        - date
        - description
      properties:
        date:
          type: string
          format: date
          example: '2026-09-29'
        categoryId:
          type: string
        description:
          type: string
        hours:
          type: number
          minimum: 0
          exclusiveMinimum: true
          maximum: 24
          multipleOf: 0.5
        projectId:
          type: string
    WorkRecord:
      type: object
      description: Optional fields are empty ("" / 0) when not set. Records always have categoryId and hours; todos may not.
      required:
        - id
        - account
        - date
        - categoryId
        - description
        - hours
        - projectId
        - createdAt
      properties:
        id:
          type: string
        account:
          type: string
        date:
          type: string
          format: date
        categoryId:
          type: string
        description:
          type: string
        hours:
          type: number
        projectId:
          type: string
        createdAt:
          type: string
          format: date-time
    Todo:
      $ref: '#/components/schemas/WorkRecord'
    WorkRecordListResponse:
      type: object
      required:
        - message
        - records
        - total
        - totalHours
      properties:
        message:
          type: string
        records:
          type: array
          items:
            $ref: '#/components/schemas/WorkRecord'
        total:
          type: integer
        totalHours:
          type: number
    WorkRecordResponse:
      type: object
      required:
        - message
        - record
      properties:
        message:
          type: string
        record:
          $ref: '#/components/schemas/WorkRecord'
    TodosResponse:
      type: object
      required:
        - message
        - todos
      properties:
        message:
          type: string
        todos:
          type: array
          items:
            $ref: '#/components/schemas/Todo'
    TodoResponse:
      type: object
      required:
        - message
        - todo
      properties:
        message:
          type: string
        todo:
          $ref: '#/components/schemas/Todo'
    CompleteTodoRequest:
      type: object
      description: categoryId, hours and projectId fill or override the todo's; the resulting record needs a category and hours.
      required:
        - date
      properties:
        date:
          type: string
          format: date
        categoryId:
          type: string
        hours:
          type: number
          minimum: 0
          exclusiveMinimum: true
          maximum: 24
          multipleOf: 0.5
        projectId:
          type: string
    WorkMember:
      type: object
      required:
        - account
        - name
      properties:
        account:
          type: string
        name:
          type: string
    WorkMembersResponse:
      type: object
      required:
        - message
        - members
      properties:
        message:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/WorkMember'
...
    WeekDay:
      type: object
      required:
        - date
        - workday
        - loggedHours
        - requiredHours
      properties:
        date:
          type: string
          format: date
        workday:
          type: boolean
        holidayName:
          type: string
          description: Name of the holiday or makeup workday on this date.
        loggedHours:
          type: number
        requiredHours:
          type: number
    WeekSummaryResponse:
      type: object
      required:
        - message
        - from
        - to
        - recordCount
        - loggedHours
        - requiredHours
        - remainingHours
        - daysOff
        - days
      properties:
        message:
          type: string
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        recordCount:
          type: integer
        loggedHours:
          type: number
        requiredHours:
          type: number
        remainingHours:
          type: number
          description: requiredHours - loggedHours, never below 0.
        daysOff:
          type: integer
          description: Weekdays off in the week.
        days:
          type: array
          items:
            $ref: '#/components/schemas/WeekDay'
    MessageResponse:
      type: object
      properties:
        message:
          type: string